feat: Explore US Bikeshare Data
Using Python to explore data related to bike share systems for three major cities in the United States — Chicago, New York City, and Washington. The code is created to import the data and answer interesting questions about it by computing descriptive statistics as well as take in raw input to create an interactive experience in the terminal to present these statistics.